stride

Trait StrideIndex

pub unsafe trait StrideIndex<T: ?Sized>: Sealed {
    type Output: ?Sized;

    // Required methods
    fn get(self, stride: &T) -> Option<&Self::Output>;
    fn get_mut(self, stride: &mut T) -> Option<&mut Self::Output>;
    unsafe fn get_unchecked(self, stride: *const T) -> *const Self::Output;
    unsafe fn get_unchecked_mut(self, stride: *mut T) -> *mut Self::Output;
    fn index(self, stride: &T) -> &Self::Output;
    fn index_mut(self, stride: &mut T) -> &mut Self::Output;
}

A helper trait used for indexing operations.

This is the Stride version of SliceIndex. You should not use or implement this trait directly but instead use the corresponding methods on Stride.

Safety

Implementations of this trait have to promise that if the argument to get_(mut_)unchecked is a safe reference, then so is the result.

Required Associated Types

type Output: ?Sized

The output type returned by methods.

Required Methods

fn get(self, stride: &T) -> Option<&Self::Output>

Returns a shared reference to the output at this location, if in bounds.

fn get_mut(self, stride: &mut T) -> Option<&mut Self::Output>

Returns a mutable reference to the output at this location, if in bounds.

unsafe fn get_unchecked(self, stride: *const T) -> *const Self::Output

Returns a shared reference to the output at this location, without performing any bounds checking.

Safety

Calling this method with an out-of-bounds index or a dangling stride pointer is undefined behavior even if the resulting reference is not used.

unsafe fn get_unchecked_mut(self, stride: *mut T) -> *mut Self::Output

Returns a mutable reference to the output at this location, without performing any bounds checking.

Safety

Calling this method with an out-of-bounds index or a dangling stride pointer is undefined behavior even if the resulting reference is not used.

fn index(self, stride: &T) -> &Self::Output

Returns a shared reference to the output at this location, panicking if out of bounds.

fn index_mut(self, stride: &mut T) -> &mut Self::Output

Returns a mutable reference to the output at this location, panicking if out of bounds.